邮件提醒天气/requirements.txt
backcall==0.1.0
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
idna==2.9
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.2.1
matplotlib==3.2.0
numpy==1.18.1
pandas==1.0.3
parso==0.6.2
pickleshare==0.7.5
prompt-toolkit==3.0.4
Pygments==2.6.1
pynput==1.6.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
